fix: reconcile worker diagnostics after main rebase
