+++ pkg/services/factory_runtime/internal/services/orchestration/runtime/worker_pool.go
@@ -867,75 +867,3 @@ func cloneSessionCapabilities(value *workers.Capabilities) *workers.Capabilities
 	clone := *value
 	return &clone
 }
-
-func recordDetachedAgentRunResponse(
-	cfg *runtimeConfig,
-	request workers.ExecuteRequest,
-	result workers.ExecuteResult,
-	executeErr error,
-) {
-	if cfg == nil || cfg.eventHistory == nil || !runtimeRequestUsesAgentRun(cfg, request) {
-		return
-	}
-	recorder, ok := cfg.eventHistory.(recordings.WorkerEventRecorder)
-	if !ok || recorder == nil {
-		return
-	}
-	dispatchID := strings.TrimSpace(request.Correlation.DispatchID)
-	if dispatchID == "" {
-		return
-	}
-
-	transcript := make([]workers.AgentRunTranscriptEntry, 0, 3)
-	appendPromptTranscriptEntry(&transcript, "system", request.Target.Prompt.SystemPrompt)
-	appendPromptTranscriptEntry(&transcript, "user", request.Target.Prompt.UserMessage)
-	appendTranscriptEntry(&transcript, "assistant", primaryOutputText(result.Output.Primary))
-	safeDiagnostics := workers.SafeWorkDiagnosticsFromWorkDiagnostics(result.Diagnostics.ToWorkDiagnostics())
-	if safeDiagnostics == nil {
-		safeDiagnostics = &workers.SafeWorkDiagnostics{}
-	}
-	agentRun := safeDiagnostics.AgentRun
-	if agentRun == nil {
-		agentRun = &workers.SafeAgentRunDiagnostic{}
-	}
-	agentRun.ExecutionBehavior = workers.AgentRunExecutionBehavior
-	agentRun.Transcript = transcript
-	safeDiagnostics.AgentRun = agentRun
-	diagnostics, err := workers.SafeWorkDiagnosticsEventPayload(safeDiagnostics)
-	if err != nil {
-		return
-	}
-	outcome := result.Outcome
-	if outcome == "" && executeErr != nil {
-		outcome = workers.ExecutionOutcomeFailed
-	}
-	eventTime := cfg.clock.Now()
-	recorder.RecordAgentRunEvent(workers.AgentRunResponseEvent{
-		ID:         fmt.Sprintf("factory-event/agent-run-response/%s", dispatchID),
-		DispatchID: dispatchID,
-		EventTime:  eventTime,
-		Tick:       detachedExecutionTick(request.Input.Dispatch.Execution),
-		Payload: workers.AgentRunResponseEventPayload{
-			AgentRunID:     fmt.Sprintf("%s/agent-run/1", dispatchID),
-			Diagnostics:    diagnostics,
-			DurationMillis: result.Metrics.Duration.Milliseconds(),
-			Outcome:        string(outcome),
-		},
-	})
-}
-
-func detachedExecutionTick(metadata work.ExecutionMetadata) int {
-	if metadata.CurrentTick != 0 {
-		return metadata.CurrentTick
-	}
-	return metadata.DispatchCreatedTick
-}
-
-func runtimeRequestUsesAgentRun(cfg *runtimeConfig, request workers.ExecuteRequest) bool {
-	lookup, ok := runtimeDefinitionLookup(cfg)
-	if !ok || lookup == nil {
-		return false
-	}
-	workstation, found := lookup.Workstation(strings.TrimSpace(request.Target.WorkstationName))
-	return found && workstation != nil && interfaces.IsAgentRunWorkstationType(workstation.Type)
-}
