+++ pkg/services/factory_runtime/internal/services/orchestration/runtime/worker_session_control_targets.go
@@ -725,6 +725,24 @@ func finalizeRuntimeWorkspaceSelection(
 	selection.workingDirectory = resolveRuntimePath(baseDirectory, selection.workingDirectory, fileSystem)
 }
 
+func runtimeWorkflowContext(cfg *runtimeConfig, sessionID string, supplied *workers.Context) *workers.Context {
+	if context := supplied.Clone(); context != nil {
+		if strings.TrimSpace(context.SessionID) == "" {
+			context.SessionID = strings.TrimSpace(sessionID)
+		}
+		return context
+	}
+	if cfg != nil {
+		if context := cfg.workflowContext.Clone(); context != nil {
+			if strings.TrimSpace(context.SessionID) == "" {
+				context.SessionID = strings.TrimSpace(sessionID)
+			}
+			return context
+		}
+	}
+	return &workers.Context{SessionID: strings.TrimSpace(sessionID)}
+}
+
 const classifierFailureRawOutputLimit = 160
 
 func normalizeScriptClassifierResult(
@@ -951,52 +969,4 @@ func recordDispatchWorkerSessionAssociation(
 		return
 	}
 	ledger.RecordDispatchWorkerSessionAssociation(tick, dispatchID, workerSessionID, requestID, eventTime)
-}
-
-func recordedObservationFromFact(fact recordedDispatchObservation, clock factory.Clock) workersessions.Observation {
-	state := fact.state
-	if state == "" {
-		state = workersessions.StateStarting
-	}
-	observation := workersessions.Observation{
-		WorkerSessionID:          fact.workerSessionID,
-		Model:                    cloneRecordedString(fact.model),
-		ReasoningEffort:          cloneRecordedString(fact.reasoningEffort),
-		ProviderSessionAvailable: fact.provider != nil && fact.provider.ID != "",
-		WorkIDs:                  append([]string(nil), fact.workIDs...),
-		TurnID:                   fact.turnID,
-		AttemptID:                fact.dispatchID,
-		State:                    state,
-		DurationBasis:            workersessions.DurationBasisUnavailable,
-		Transcript:               workersessions.TranscriptAvailabilityUnavailable,
-	}
-	if fact.provider != nil {
-		observation.ProviderSession = providerSessionRef(*fact.provider)
-	}
-	if !fact.startedAt.IsZero() {
-		started := fact.startedAt.UTC()
-		observation.StartedAt = &started
-		if fact.endedAt != nil {
-			ended := fact.endedAt.UTC()
-			observation.EndedAt = &ended
-			duration := ended.Sub(started)
-			if duration < 0 {
-				duration = 0
-			}
-			observation.Duration = &duration
-			observation.DurationBasis = workersessions.DurationBasisRecordedTimestamps
-		} else if !state.Terminal() && clock != nil {
-			duration := clock.Now().Sub(started)
-			if duration < 0 {
-				duration = 0
-			}
-			observation.Duration = &duration
-			observation.DurationBasis = workersessions.DurationBasisActiveClock
-		}
-	}
-	if fact.failure != nil {
-		failure := *fact.failure
-		observation.Failure = &failure
-	}
-	return observation
 }
